pyproject.toml
[project]
name = "pydantic-ai-production-ready-template"
version = "0.1.0"
description = "Pydantic AI Production Ready Template"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=1.13.0",
    "fastapi>=0.121.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "SQLAlchemy>=2.0.44",
    "alembic>=1.17.1",
    "asyncpg>=0.31.0",
    "greenlet>=3.2.4",
    "redis>=7.1.0",
    "fastapi-guard>=4.2.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "starlette-admin>=0.15.1",
]

[dependency-groups]
dev = [
    "cookiecutter>=2.6.0",
    "cookiecutter-data-science>=2.3.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "nox>=2025.5.1",
    "pre-commit>=4.3.0",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "pydantic-evals[logfire]>=1.16.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
